+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.2.0
+- Add palegarden -> vanillabackport
+- Add BWG skyrise -> skyris
+
 ## 0.1.0
 - Initial